requirements.txt
aiohttp==3.8.6
pandas==1.5.3
python_dateutil==2.8.2
tabulate
